all is well but death may bug on the nbr dead

+++ srcs/utils.h
@@ -6,7 +6,7 @@
 /*   By: mmidon <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/12/01 13:36:36 by mmidon            #+#    #+#             */
-/*   Updated: 2023/01/18 10:00:02 by mmidon           ###   ########.fr       */
+/*   Updated: 2023/01/20 08:31:04 by mmidon           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 #include <pthread.h>
@@ -17,7 +17,7 @@
 # define UTILS_H
 
 void    ft_print(int nbr, char *action, t_args *args);
-int		ft_time(int start);
+long long int	ft_time(int start);
 void    ft_death(t_args *args);
 
 #endif
